requirements.txt
fastapi
uvicorn
langchain
langchain_community
langchain-google-genai
python-dotenv
python-multipart